Add .babelrc for styled ssr Remove App component
