+++ applications/web/.babelrc
@@ -1,0 +1,19 @@
+{
+    "env": {
+        "development": {
+            "plugins": [
+                ["styled-components", { "ssr": true, "displayName": true, "preprocess": false } ]
+            ],
+            "presets": ["next/babel"]
+        },
+        "production": {
+            "plugins": [
+                ["styled-components", { "ssr": true, "displayName": true, "preprocess": false } ]
+            ],
+            "presets": ["next/babel"]
+        }
+    },
+    "plugins": [
+        ["styled-components", { "ssr": true, "displayName": true, "preprocess": false } ]
+    ]
+}
